# src/main/resources/api/api.yaml
openapi: 3.1.0
info:
  title: Emarket
  description: 'Một website thương mại điện tử nơi người dùng có thể mua các sản phầm mà người dùng khác bán, hoặc bán sản phẩm của chính mình.'
  version: 1.0.0
  contact:
    name: Phạm Minh Nguyên
    email: 21020084@vnu.edu.vn
servers:
  - url: localhost:8080/
    description: Máy chủ chưa triển khai.
...
paths:
...
  /api/v1/user/{username}:

    get:
      tags:
        - user
      description: Lấy thông tin về một tài khoản người dùng.
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      responses:
        200:
          description: Lấy thông tin thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags:
        - user
      description: Xóa tài khoản của người dùng khỏi hệ thống.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      responses:
        200:
          description: Xóa tài khoản thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
                description: Thông báo thao tác thành công.
  /api/v1/user/{username}/details:

    get:
      tags:
        - user
      description: Lấy thông tin chi tiết về một tài khoản người dùng.
      operationId: getDetails
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      responses:
        200:
          description: Lấy thông tin thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'

    put:
      tags:
        - user
      description: Cập nhật thông tin người dùng.
      operationId: updateDetails
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        200:
          description: Cập nhật thông tin thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdate'
  /api/v1/user/{username}/details/password:

    put:
      tags:
        - user
      description: Thay đổi mật khẩu của tài khoản.
      operationId: updatePassword
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        200:
          description: Thay đổi mật khẩu thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
                description: Thông báo thao tác thành công.
  /api/v1/user/{username}/details/avatar:

    post:
      tags:
        - user
      description: Cập nhật ảnh đại diện của tài khoản.
      operationId: updateAvatar
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Tệp ảnh đại diện được tải lên.
      responses:
        200:
          description: Cập nhật ảnh đại diện thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
                description: Thông báo thao tác thành công.

    delete:
      tags:
        - user
      description: Xóa ảnh đại diện của người dùng.
      operationId: deleteAvatar
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: tên đâng nhập của tài khoản.
      responses:
        200:
          description: Xóa ảnh đại diện thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
                description: Thông báo thao tác thành công.
...
  /api/v1/products:

    post:
      tags:
        - product
      description: Tạo một sản phẩm mới.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Tạo sản phẩm thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
                description: Sản phẩm mới được tạo thành công.
  /api/v1/product/{productId}:

    get:
      tags:
        - product
      description: Lấy thông tin về một sản phẩm.
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: int
            description: Id của sản phầm..
      responses:
        200:
          description: Lấy thông tin thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
                description: Thông tin về sản phẩm.

    put:
      tags:
        - product
      description: Cập nhật thông tin sản phẩm.
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: int
            description: Id của sản phầm.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        200:
          description: Cập nhật thông tin thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags:
        - product
      description: Xóa sản phẩm khỏi hệ thống.
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: int
            description: Id của sản phầm.
      responses:
        200:
          description: Xóa sản phẩm thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
                description: Thông báo thao tác thành công.
  /api/v1/product/{productId}/images:

    post:
      tags:
        - product
      description: Cập nhật ảnh của sản phẩm.
      operationId: updateProductImage
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: int
            description: Id của sản phẩm.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                    description: Tệp ảnh được tải lên.
      responses:
        200:
          description: Cập nhật ảnh thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagesUpdateResponse'
                description: Thông báo thao tác thành công.
  /api/v1/category/{categoryName}/products:

    get:
      tags:
        - product
      operationId: getProductsByCategory
      parameters:
        - name: categoryName
          in: path
          required: true
          schema:
            type: string
            description: Tên của loại sản phầm.
        - name: page
          in: query
          required: true
          schema:
            type: int
            description: Số trang.
        - name: size
          in: query
          required: true
          schema:
            type: int
            description: Số sản phẩm trong một trang.
        - name: sortBy
          in: query
          required: true
          schema:
            type: string
            description: Tiêu chí sắp xếp sản phẩm.
            enum:
              - name
              - unitPrice
              - updatedAt
        - name: sortDirection
          in: query
          required: true
          schema:
            type: string
            description: Hướng sắp xếp sản phẩm.
            enum:
              - asc
              - desc
      responses:
        200:
          description: Lấy thông tin thành công.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/v1/user/{username}/products:

    get:
      tags:
        - product
      operationId: getProductsBySeller
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            description: Tên đăng nhập của tài khoản bán sản phẩm đó.
        - name: page
          in: query
          required: true
          schema:
            type: int
            description: Số trang.
        - name: size
          in: query
          required: true
          schema:
            type: int
            description: Số sản phẩm trong một trang.
        - name: sortBy
          in: query
          required: true
          schema:
            type: string
            description: Tiêu chí sắp xếp sản phẩm.
            enum:
              - name
              - unitPrice
              - updatedAt
        - name: sortDirection
          in: query
          required: true
          schema:
            type: string
            description: Hướng sắp xếp sản phẩm.
            enum:
              - asc
              - desc
      responses:
        200:
          description: Lấy thông tin thành công.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
components:
  schemas:
...
    SuccessResponse:
      type: object
      description: Nội dung phản hồi của một yêu cầu thành công.
      required:
        - text
      properties:
        text:
          type: string
          description: Mô tả dễ hiểu về thao tác thành công.
    ImagesUpdateResponse:
      type: object
      description: Nội dung phản hồi của một yêu cầu cập nhật ảnh thành công.
      required:
        - imageUrls
      properties:
        imageUrls:
          type: array
          description: Danh sách các đường dẫn đến file ảnh của sản phẩm.
          items:
            type: string
...
    UpdatePasswordRequest:
        type: object
        description: Thông tin cần thiết để thay đổi mật khẩu của một người dùng.
        required:
            - oldPassword
            - newPassword
        properties:
          oldPassword:
            type: string
            minLength: 8
            maxLength: 60
            format: password
            description: Mật khẩu cũ của người dùng.
          newPassword:
            $ref: '#/components/schemas/Password'
            description: Mật khẩu mới của người dùng.
    Password:
      type: object
      description: Thông tin về mật khẩu của một tài khoản người dùng.
      required:
        - text
        - updatedAt
      properties:
        text:
          type: string
          minLength: 8
          maxLength: 60
          format: password
          description: Nội dung của mật khẩu.
        updatedAt:
          type: string
          format: date-time
          description: Thời điểm gần đây nhất mật khẩu đó được thay đổi.
...
    Profile:
      allOf:
        - $ref: '#/components/schemas/UserUpdate'
        - type: object
          description: Thông tin chi tiết hơn về User.
          required:
            - age
            - authority
          properties:
            authority:
              type: string
              description: Vai trò của người dùng đó trong hệ thống.
              enum:
                - ROLE_ADMIN
                - ROLE_USER
    User:
      type: object
      description: Thông tin về một người dùng trong hệ thống.
      required:
        - id
        - username
      properties:
        id:
          type: int
          description: Id của tài khoản.
        imageUrl:
          type: string
          description: Đường dẫn đến file ảnh đại diện của tài khoản.
        username:
          type: string
          minLength: 6
          maxLength: 36
          description: Tên đăng nhập của tài khoản.
        profile:
          $ref: '#/components/schemas/Profile'
          description: Thông tin chi tiết về tài khoản của người dùng đó.
    UserUpdate:
      type: object
      description: Dữ liệu của User trong thao tác UPDATE.
      properties:
        firstName:
          type: string
          maxLength: 36
          description: Tên của chủ tài khoản.
        lastName:
          type: string
          maxLength: 36
          description: Họ của chủ tài khoản.
        email:
          type: string
          maxLength: 100
          format: email
          description: Địa chỉ email của chủ tài khoản.
        age:
          type: int
          minimum: 13
          description: Tuổi của chủ tài khoản.
        gender:
          type: string
          description: Giới tính của chủ tài khoản.
          enum:
            - FEMALE
            - MALE
            - OTHER
            - UNKNOWN
        address:
          $ref: '#/components/schemas/Address'
          description: Địa chỉ của người dùng đó.
    Category:
      type: object
      description: Thông tin về một loại sản phẩm.
      required:
        - name
        - parentName
      properties:
        id:
          type: int
          description: Id của loại sản phẩm.
        name:
          type: string
          maxLength: 50
          description: Tên của loại sản phẩm.
        level:
          type: int
          description: Cấp độ của loại sản phẩm trong cây phân cấp loại sản phẩm.
        path:
          type: string
          description: Đường dẫn đến loại sản phẩm trong cây phân cấp loại sản phẩm.
        parentName:
          type: string
          description: Tên của loại sản phẩm cha.
        parent:
          $ref: '#/components/schemas/Category'
          description: Loại sản phẩm cha của loại sản phẩm này.
    Product:
      type: object
      description: Thông tin chi tiết hơn về Product.
      properties:
        id:
          type: int
          description: Id của sản phầm.
        name:
          type: string
          maxLength: 50
          description: Tên của sản phầm.
        coverImageUrl:
          type: string
          description: Đường dẫn đến file ảnh bìa của sản phẩm.
        unitPrice:
          type: int
          minimum: 0
          description: Đơn giá của sản phẩm.
        updatedAt:
          type: string
          format: date-time
          description: Thời điểm gần nhất thông tin sản phẩm được cập nhật.
        shortDescription:
          type: string
          maxLength: 100
          description: Mô tả ngắn gọn của sản phầm.
        description:
          type: string
          maxLength: 200
          description: Mô tả chi tiết của sản phầm.
        qtyInStock:
          type: int
          minimum: 0
          description: Số lượng sẵn bán của sản phẩm.
        imageUrls:
          type: array
          description: Danh sách các đường dẫn đến file ảnh của sản phẩm.
          items:
            type: string
            description: Đường dẫn đến file ảnh của sản phẩm.
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
            description: Những loại sản phẩm mà sản phẩm này thuộc về.
        sellerUsername:
          type: String
          description: Tên đăng nhập của người bán sản phẩm đó.
        seller:
          $ref: '#/components/schemas/User'
          description: Người bán sản phẩm đó.
...
    Address:
      type: object
      description: Thông tin về địa chỉ của một người dùng.
      required:
        - province
        - district
        - streetAndNumber
      properties:
        province:
          type: string
          description: Tên tỉnh/thành phố.
        district:
          type: string
          description: Tên xã/phường/thị trấn.
        streetAndNumber:
          type: string
          description: Tên đường/tuyến phố và số nhà.